pac1-coder/pyproject.toml
[project]
name = "bitgn-pac1-coder"
version = "0.1.0"
description = "Runnable Python sample for the BitGN PAC1 benchmark"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "connect-python>=0.8.1",
    "protobuf>=4.25.0",
    "httpx>=0.27.0",
    "openai>=2.26.0",
    "pydantic>=2.12.5",
    "annotated-types>=0.7.0",
    "anthropic>=0.86.0",
    "dspy-ai>=2.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
# AICODE-NOTE: Uses locally generated protobuf files (pac1-coder/bitgn/) and
# connect-python instead of external buf.build SDK packages, mirroring
# the sandbox-py approach for offline/authenticated-free operation.
package = false

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
